Fix #242 Meson tests
