+++ examples/output_kind/Makefile.meson
@@ -1,6 +1,7 @@
 include ../make.meson.inc
 
 NAME     := pywrapper
+WRAPFLAGS += --type-check --kind-map kind.map
 
 test: build
 	$(PYTHON) test.py
