+++ examples/return_array/Makefile.meson
@@ -3,4 +3,4 @@ include ../make.meson.inc
 NAME     := pywrapper
 
 test: build
-	$(PYTHON) tests.py
+	$(PYTHON) test.py
