+++ examples/type_check/Makefile.meson
@@ -1,7 +1,7 @@
 include ../make.meson.inc
 
 NAME     := pywrapper
-WRAPFLAGS += --type-check
+WRAPFLAGS += --type-check --kind-map kind.map
 
 test: build
 	$(PYTHON) type_check_test.py
